+++ src/main/java/onlyoffice/managers/document/DocumentManager.java
@@ -7,6 +7,7 @@ public interface DocumentManager extends Serializable {
     public String getKeyOfFile(Long attachmentId);
     public String createHash(String str);
     public String readHash(String base64);
+    public String getCorrectName(String fileName, String fileExt, Long pageID);
     public Long createDemo(String fileName, String fileExt, Long pageID);
     public String getDocType(String ext);
     public String getEditorType (String userAgent);