+++ src/main/java/onlyoffice/managers/url/UrlManagerImpl.java
@@ -46,6 +46,7 @@ public class UrlManagerImpl implements UrlManager {
     private final String callbackServlet = "plugins/servlet/onlyoffice/save";
     private final String historyServlet = "plugins/servlet/onlyoffice/history";
     private final String fileProviderServlet = "plugins/servlet/onlyoffice/file-provider";
+    private final String APIServlet = "plugins/servlet/onlyoffice/api";
 
     @ComponentImport
     private final PluginSettingsFactory pluginSettingsFactory;
@@ -89,10 +90,10 @@ public class UrlManagerImpl implements UrlManager {
     public String getFileUri(Long attachmentId) {
         String hash = documentManager.createHash(Long.toString(attachmentId));
 
-        String callbackUrl = getFileProviderUri() + "?vkey=" + GeneralUtil.urlEncode(hash);
-        log.info("fileUrl " + callbackUrl);
+        String fileUri = getConfluenceBaseUrl() + fileProviderServlet + "?vkey=" + GeneralUtil.urlEncode(hash);
+        log.info("fileUrl " + fileUri);
 
-        return callbackUrl;
+        return fileUri;
     }
 
     public String getAttachmentDiffUri(Long attachmentId) {
@@ -116,10 +117,16 @@ public class UrlManagerImpl implements UrlManager {
         return historyDataUri;
     }
 
-    public String getFileProviderUri() {
-        String fileProviderUri = getConfluenceBaseUrl() + fileProviderServlet;
+    public String getAttachmentDataUri() {
+        String attachmentDataUri = getConfluenceBaseUrl() + APIServlet + "?type=attachment-data";
 
-        return fileProviderUri;
+        return attachmentDataUri;
+    }
+
+    public String getSaveAsUri() {
+        String saveAsUri = getConfluenceBaseUrl() + APIServlet + "?type=save-as";
+
+        return saveAsUri;
     }
 
     public String getCallbackUrl(Long attachmentId) {