+++ src/main/java/onlyoffice/utils/attachment/AttachmentUtil.java
@@ -12,6 +12,7 @@ import java.util.List;
 public interface AttachmentUtil extends Serializable {
     public boolean checkAccess(Long attachmentId, User user, boolean forEdit);
     public boolean checkAccess(Attachment attachment, User user, boolean forEdit);
+    public boolean checkAccessCreate(User user, Long pageId);
     public void saveAttachmentAsNewVersion(Long attachmentId, InputStream attachmentData, int size, ConfluenceUser user)
             throws IOException, IllegalArgumentException;
     public void updateAttachment(Long attachmentId, InputStream attachmentData, int size, ConfluenceUser user);
@@ -31,7 +32,9 @@ public interface AttachmentUtil extends Serializable {
     public int getVersion (Long attachmentId);
     public Attachment getAttachmentChanges (Long attachmentId);
     public Attachment getAttachmentDiff (Long attachmentId);
+    public String getAttachmentPageTitle (Long attachmentId);
     public Long getAttachmentPageId (Long attachmentId);
     public String getAttachmentSpaceName (Long attachmentId);
     public String getAttachmentSpaceKey (Long attachmentId);
+    public Attachment createNewAttachment (String title, String mimeType, InputStream file, int size, Long pageId, ConfluenceUser user) throws IOException;
 }