+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ## Added
+- "save as" attachment on page from editor
 - open non-editable formats on views
 - formats for conversion docm, dot, dotx, epub, htm, html, otp, ots, ott, pot, potm, potx, pps, ppsm, ppsx,
  pptm, rtf, xlsm, xlt, xltm, xltx